+++ kubernetes/utils.go
@@ -150,11 +150,6 @@ var (
 	INGRESSNAME            = fmt.Sprintf("%s-ingress", version.Name)
 )
 
-type K8sWorkloadResult struct {
-	Result interface{} `json:"result,omitempty"`
-	Error  interface{} `json:"error,omitempty"`
-}
-
 type K8sNewWorkload struct {
 	Name        string `json:"name"`
 	YamlString  string `json:"yamlString"`
@@ -200,19 +195,19 @@ func WorkloadsForAccesslevel(access dtos.AccessLevel) []string {
 	return resources
 }
 
-func WorkloadResult(result interface{}, err interface{}) K8sWorkloadResult {
+func WorkloadResult(result interface{}, err interface{}) utils.K8sWorkloadResult {
 	fmt.Println(reflect.TypeOf(err))
 	if fmt.Sprint(reflect.TypeOf(err)) == "*errors.errorString" {
 		err = err.(error).Error()
 	}
-	return K8sWorkloadResult{
+	return utils.K8sWorkloadResult{
 		Result: result,
 		Error:  err,
 	}
 }
 
-func WorkloadResultError(error string) K8sWorkloadResult {
-	return K8sWorkloadResult{
+func WorkloadResultError(error string) utils.K8sWorkloadResult {
+	return utils.K8sWorkloadResult{
 		Result: nil,
 		Error:  error,
 	}